refactor(ui): move Modal and MonacoDiff into components/primitives
Both are generic — a dialog shell and a Monaco diff viewer — and neither
contains a single reference to operate, cluster, inspect, dsl or live-debug.
They lived under `features/operate/_shared/` because that is where they were
first written, and fifteen files came to import them from there: `shell/`,
`render/widgets/`, `layer/`, and three features besides operate. A shell-level
prompt reaching into a feature folder for a dialog is the shape that says the
address is wrong rather than the rule.

Moving them takes the cross-feature imports the layering rule forbids from
fourteen down to one — the admin editor for the 3D map reading that feature's
own config composable, which is a page editing the thing it is the editor for.

No behaviour change: the two files are unmodified apart from a relative
`monaco/setup` import that the new depth broke, now on the `@/` alias like
everything around it.
