+++ apps/ui/CLAUDE.md
@@ -85,6 +85,8 @@ Each control owns its own reset and its own indicator, and resolves as its query
 
 `api/` is the only path to HTTP. `shell/` knows about layers and routes, never about a feature's data. `controls/` owns the time range and the refresh ticker; pages subscribe and never own them. `state/` is global Pinia that survives navigation. `features/<feature>/` and `layer/<tab>/` are self-contained — views, composables and components together. `render/` is template-driven and generic. `components/{primitives,charts,icons}/` is feature-AGNOSTIC: the moment a shared component needs feature data, it moves into the feature rather than the shared pile growing a special case.
 
-**Features do not import from each other**, and the exceptions today say what the rule is for rather than weakening it. `Modal` and `MonacoDiff` sit under `features/operate/_shared/` and are borrowed by admin and events — they are primitives in everything but their address, and moving them is a tidy-up nobody has done. The admin editor for the 3D map reads that feature's own config composable, which is a page editing the thing it is the editor for. Neither is one feature reaching into another's *data*, which is what the rule exists to stop.
+**Features do not import from each other.** The one exception today is the admin editor for the 3D map, which reads that feature's own config composable — a page editing the thing it is the editor for, not one feature reaching into another's *data*, which is what the rule exists to stop.
+
+A component that several features borrow is a primitive whichever folder it happens to sit in, and belongs in `components/primitives/`. `Modal` and `MonacoDiff` lived under `features/operate/_shared/` for a while because that is where they were first written; fifteen files across `shell/`, `render/`, `layer/` and three features imported them from there, which is the shape that tells you the address was wrong rather than the rule.
 
 **A chart component owns its ECharts instance; a VIEW never creates one.** Most go through the shared wrappers in `components/charts/`; a couple of chart types specific to one feature have their own component beside that feature, which is the same rule applied to a component that knows a feature's shape. What is banned is a view calling `echarts.init` inline, because then nothing owns disposal or theming. D3 lifecycles belong to a composable that tears down on unmount.
