+++ apps/ui/src/features/admin/overview-templates/OverviewTemplatesAdmin.vue
@@ -54,8 +54,8 @@ import { refreshConfigBundle } from '@/controls/configBundle';
 import { stableStringify } from '@/utils/stableJson';
 import TemplatePicker, { type TemplatePickerEntry } from '@/features/admin/_shared/TemplatePicker.vue';
 import { useTemplateSync } from '@/features/admin/_shared/useTemplateSync';
-import Modal from '@/features/operate/_shared/Modal.vue';
-import MonacoDiff from '@/features/operate/_shared/MonacoDiff.vue';
+import Modal from '@/components/primitives/Modal.vue';
+import MonacoDiff from '@/components/primitives/MonacoDiff.vue';
 import OverviewEditCanvas from './OverviewEditCanvas.vue';
 import OverviewWidgetDrawer from './OverviewWidgetDrawer.vue';
 import NewOverviewDashboardModal from './NewOverviewDashboardModal.vue';
